scripts/requirements.txt
python-amazon-paapi>=6.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
